fix replace media action
