+++ src/Providers/EventServiceProvider.php
@@ -7,7 +7,6 @@ use FluxErp\Listeners\Auth\LogoutListener;
 use FluxErp\Listeners\BroadcastEventSubscriber;
 use FluxErp\Listeners\CacheKeyWrittenListener;
 use FluxErp\Listeners\MailMessage\CreateMailExecutedSubscriber;
-use FluxErp\Listeners\Media\ReplaceMediaSubscriber;
 use FluxErp\Listeners\MessageSendingEventSubscriber;
 use FluxErp\Listeners\NotificationEloquentEventSubscriber;
 use FluxErp\Listeners\Order\OrderInvoiceAddedSubscriber;
@@ -62,7 +61,6 @@ class EventServiceProvider extends ServiceProvider
         OrderInvoiceAddedSubscriber::class,
         MessageSendingEventSubscriber::class,
         CreateMailExecutedSubscriber::class,
-        ReplaceMediaSubscriber::class,
     ];
 
     public function boot(): void
